email_automation/requirements.txt
python-dotenv==1.0.1
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
pydantic==2.9.2
